requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
openai
huggingface
pypdf
python-dotenv
ipykernel
huggingface_hub
numpy==1.23.4
llama-cpp-python==0.1.78
flask
-e .